updating the code to add functions
