+++ Output_csv/GST Collection Domestic.csv
@@ -1,0 +1,67 @@
+,industry_id,GST Collection Domestic
+0,Agriculture,2936.6839151596505
+1,Livestock,215.71425479473936
+2,Forestry and Logging,1044.0988889389785
+3,Fishing & Aquaculture,4.741831411643274
+4,Coal & Lignite,1339.3306534638818
+5,Crude Petroleum,2397.040260977175
+6,Natural Gas,832.5162091600454
+7,Iron Ores,1245.746231295877
+8,Non ferrous metal ores,407.9121220816589
+9,Other Mining,14646.819837901532
+10,"Production, processing and preservation of meat, fish, fruit, vegetables, oils and fats",3139.7804191267633
+11,Manufacture of dairy products,4610.104098550463
+12,"Manufacture of grain mill products, etc. and animal feeds",7687.217861153145
+13,Manufacture of other food products,16402.604374149905
+14,Manufacture of beverages,1367.6559625603684
+15,Manufacture of tobacco products,35564.515251803794
+16,Manufacture of textiles + cotton ginning,11248.967772988184
+17,"Manufacture of wearing apparel, except custom tailoring",2489.8602112207186
+18,Manufacture of leather and related products,947.2158094758879
+19,Manufacture of Basic Iron and Steel + Casting of iron and steel,48568.461979570035
+20,Manufacture of basic precious and non-ferrous metals + Casting of non-ferrous metals,6029.843275496089
+21,"Manufacture of fabricated metal products, except machinery and equipments",7475.38904011892
+22,"Manufacture of electronic component, consumer electronics, magnetic and optical media",4773.682289263408
+23,Manufacture of computer and peripheral equipment,1955.4114626297614
+24,Manufacture of communication equipments,1224.9638366390202
+25,Manufacture of optical and electronics products n.e.c,2004.6837702338255
+26,Manufacture of Electrical equipments,18431.879807970305
+27,Manufacture of machinery and equipments n.e.c,29338.71957697018
+28,Manufacture of Transport,53980.869033786854
+29,Manufacture of coke and refined petroleum products,21704.04541733159
+30," Manufacture of chemical and chemical products except pharmaceuticals, medicinal and botanical products",38090.58912513987
+31,Manufacture of pharmaceutical; medicinal chemicals and botanical products,10687.618610347106
+32,Manufacture of rubber & plastic products,19443.04232626972
+33,Manufacture of other non-metallic mineral products,31645.26170565917
+34,"Manufacture of wood and of products of wood and cork, except furniture; manufacture of articles of straw and plaiting material",363.6315840542733
+35,Manufacture of paper and paper products,5435.007264895326
+36,Printing and reproduction of recorded media except publishing,2043.0200729676906
+37,Manufacture of furniture,1611.2480993218724
+38,Other Manufacturing,6333.785999068356
+39,Repair and installation of machinery and equipments,479.86794409113077
+40,Construction,52713.90473761426
+41,Electricity,17270.002815933636
+42,Gas,1810.1242077159689
+43,Water supply,1402.9974670751742
+44,Railway Transport,2540.3429727828147
+45,Land Transport,15239.859053236745
+46,Water Transport,460.17148982770436
+47,Air Transport,1401.4297744888565
+48,Supportive & Auxilliary transport activities,3293.6990357624913
+49,Storage & warehousing,289.99817850521623
+50,Communication,8890.438895128862
+51,Trade,28778.617424655244
+52,Hotels & Restaurant,6631.11707154799
+53,Financial Services,14025.365385869205
+54,Insurance Services,2696.3198890795866
+55,Ownership of dwellings,0.0
+56,Education & Research,8844.588067679968
+57,Medical and Health,4568.403143828541
+58,Legal Services,1130.9457398359548
+59,Computer related services,9307.390988084138
+60,Other Business services,7219.434051678614
+61,Real estate activities,3173.85017171337
+62,Renting of machinery and equipment,619.424550293358
+63,"Community, Social & personal services",497.12200589112854
+64,Other services,2550.618694136603
+65,Public admn. & defence,0.0
