+++ Output_csv/itc_comm.csv
@@ -1,0 +1,140 @@
+4.664911413185976699e+02
+3.099069749350519487e+02
+1.573888674507952601e+02
+1.701695623747026431e+01
+7.109535401137538457e+00
+1.405310794332833346e+01
+2.049756426158771205e+02
+2.825279806746783038e+02
+5.444611757631521414e+02
+8.737247091513012265e+02
+3.554091552828344991e+00
+8.364011152732383607e+02
+2.357353925745956644e+02
+1.085920240325283658e+03
+1.149401858771043550e+02
+9.056936419164743768e+01
+7.198188949171941431e+02
+8.129807400229061614e+02
+5.691371818855485571e+02
+2.023409121081596140e+03
+4.556069933317468781e+02
+3.920232466826203410e-01
+1.015335520103370612e+02
+2.280930026500477084e+02
+1.453550662732532828e+03
+0.000000000000000000e+00
+1.879343412321137976e+03
+2.879028738097715490e+01
+2.084296524898740444e+01
+3.222350798907912576e+03
+2.002985046860949069e+03
+5.767137920719985232e+03
+2.469167898595783754e+03
+9.655833515160814784e+01
+6.013782226785459528e+01
+4.730779954036818680e+01
+6.045102427936627691e+02
+1.657210224280211378e+03
+1.840586345796510370e+00
+1.893051623818757071e+04
+3.037000646441934464e+02
+9.295012834746918529e+00
+4.838608122213126705e+01
+1.619298745027650511e+04
+1.128463137014688618e+04
+1.405586800661725101e+04
+1.360085357805241074e+03
+1.025360891612730484e+04
+3.136330581648693624e+03
+5.042203075155553961e+04
+1.795795346394069384e+03
+5.271204296880222273e+02
+2.554685718815004748e+03
+3.892343908152280392e+02
+6.957001526422808820e+04
+3.248635646149035165e+04
+3.419780289335040288e+04
+5.895100582621823833e+02
+9.428778391504330330e+02
+7.446680078017483311e+02
+2.192079975583535997e+04
+1.270697632220057858e+04
+2.770456225862571046e+03
+1.480152365455295012e+03
+9.225170364640942466e+01
+2.983364362457426250e+04
+1.327169685490735901e+03
+9.210955690452876297e+03
+6.243606760743645282e+04
+6.946281398700865975e+04
+4.035866349542292301e+04
+2.284752781350029181e+03
+1.565454862465796396e+04
+4.078164167812107917e+04
+8.413116914618221927e+03
+4.135705116465919673e+03
+4.763573418865600252e+04
+6.076975890305591747e+04
+5.809427595204459794e+04
+3.957276810561833554e+04
+2.347977612022255198e+04
+3.116371277358613224e+04
+3.837367984327293379e+04
+2.240273055841303722e+05
+5.951809203869943303e+03
+4.043868497122209374e+04
+6.724037087117174815e+04
+6.136662448662415954e+03
+6.747176308666244950e+03
+5.263748467408911529e+04
+4.687484571291870452e+03
+1.285891773377835307e+04
+2.109344236861337413e+04
+1.154155563700474013e+05
+6.262722460485118791e+04
+2.249323843147266598e+04
+1.501084917225894606e+04
+3.043937551292141143e+04
+3.442992290290345863e+03
+1.201885269202364179e+04
+7.254453087082198181e+04
+2.037603639521241348e+04
+4.075240999635037952e+03
+7.041218271756315517e+02
+8.075345841042047141e+03
+2.280147448953771091e+05
+6.140404808896659233e+04
+1.515539130009895598e+04
+7.826435092685885593e+02
+1.196735160633887426e+04
+4.125653456983684009e+04
+9.847119726528224419e+03
+7.990223057462932775e+04
+3.292605981140715448e+04
+2.743367582595729345e+03
+2.290791310009043173e+03
+4.726107320047818212e+04
+4.422947915683201359e+02
+1.001494145824123189e+04
+3.836666715693388596e+03
+2.301667948299155250e+04
+2.116572064390604737e+03
+6.949945961820072853e+02
+4.974456440493296213e+03
+4.379827334353392985e+02
+1.342718340094111409e+04
+2.118243604423844590e+04
+4.072237837224576197e+03
+0.000000000000000000e+00
+4.988318991845547316e+03
+9.355136611299815286e+02
+1.689468390037056906e+03
+1.708061440432440349e+03
+9.214005128688824698e+03
+1.405690396789581973e+04
+0.000000000000000000e+00
+1.335793513586630979e+04
+6.899635393167243819e+03
+7.508007674852998434e+02
+3.852186083952010904e+03
