+++ Output_csv/rate_vec.csv
@@ -1,0 +1,140 @@
+0.000000000000000000e+00
+1.750000000000000167e-02
+1.000000000000000194e-02
+5.349999999999999895e-02
+0.000000000000000000e+00
+1.600000000000000033e-02
+2.500000000000000486e-03
+5.000000000000000278e-02
+2.500000000000000139e-02
+5.000000000000000971e-02
+5.500000000000000028e-02
+0.000000000000000000e+00
+8.500000000000000611e-03
+1.830000000000000515e-01
+2.500000000000000139e-02
+5.000000000000000278e-02
+1.389999999999999569e-01
+3.200000000000000067e-02
+2.400000000000000050e-02
+0.000000000000000000e+00
+4.749999999999999362e-02
+3.500000000000000333e-02
+4.400000000000000439e-02
+5.000000000000000278e-02
+1.199999999999999956e-01
+0.000000000000000000e+00
+5.000000000000000278e-02
+0.000000000000000000e+00
+2.400000000000000050e-02
+1.160000000000000198e-01
+5.000000000000000278e-02
+0.000000000000000000e+00
+1.149999999999999911e-01
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+1.199999999999999956e-01
+5.000000000000000278e-02
+5.000000000000000278e-02
+1.199999999999999956e-01
+5.000000000000000278e-02
+1.199999999999999956e-01
+6.750000000000000444e-02
+2.500000000000000139e-02
+1.799999999999999933e-01
+0.000000000000000000e+00
+1.799999999999999933e-01
+5.000000000000000278e-02
+5.000000000000000278e-02
+2.800000000000000266e-01
+8.499999999999999223e-02
+1.499999999999999944e-01
+1.000000000000000056e-01
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.190000000000000085e-01
+1.114999999999999880e-01
+1.799999999999999933e-01
+1.542857142857142760e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+5.000000000000000278e-02
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+5.000000000000000278e-02
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.199999999999999956e-01
+1.799999999999999933e-01
+5.000000000000000278e-02
+1.799999999999999933e-01
+1.920000000000000040e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+2.800000000000000266e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+2.800000000000000266e-01
+2.800000000000000266e-01
+1.199999999999999956e-01
+2.800000000000000266e-01
+2.800000000000000266e-01
+1.199999999999999956e-01
+1.799999999999999933e-01
+5.000000000000000278e-02
+5.000000000000000278e-02
+2.800000000000000266e-01
+2.800000000000000266e-01
+1.199999999999999956e-01
+5.000000000000000278e-02
+1.799999999999999933e-01
+2.999999999999999889e-02
+1.799999999999999933e-01
+1.799999999999999933e-01
+0.000000000000000000e+00
+5.000000000000000278e-02
+0.000000000000000000e+00
+1.799999999999999933e-01
+1.799999999999999933e-01
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+5.000000000000000278e-02
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+1.799999999999999933e-01
+0.000000000000000000e+00
+1.799999999999999933e-01
+1.799999999999999933e-01
+0.000000000000000000e+00
+1.799999999999999933e-01
